Corrigindo erros
