+++ packages/web/components/Hero/index.js
@@ -1,20 +1,2 @@
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-import { HeroImage, Content } from './styles';
-import HeroSearch from './HeroSearch';
-
-const Hero = () => {
-	const { t } = useTranslation(['common']);
-
-	return (
-		<HeroImage>
-			<Content>
-				<h1>{t('common:heroTitle')}</h1>
-				<p>{t('common:heroSubTitle')}</p>
-				<HeroSearch />
-			</Content>
-		</HeroImage>
-	);
-};
-
-export default Hero;
+/* eslint-disable import/prefer-default-export */
+export { default as Hero } from './Hero';
