+++ packages/web/components/Modal/index.js
@@ -1,1 +1,2 @@
-export * from './ModalProvider';
+/* eslint-disable import/prefer-default-export */
+export { default as ModalProvider } from './ModalProvider';
