+++ packages/web/components/User/index.js
@@ -1,1 +1,2 @@
-export * from './UserProvider';
+/* eslint-disable import/prefer-default-export */
+export { default as UserProvider } from './UserProvider';
